Extract function into own file
