+++ packages/media/src/index.js
@@ -45,3 +45,4 @@ export { default as resourceList } from './resourceList';
 export { default as isAnimatedGif } from './isAnimatedGif';
 export { default as hasVideoGotAudio } from './hasVideoGotAudio';
 export { default as canTranscodeResource } from './canTranscodeResource';
+export { default as getImageFromVideo } from './getImageFromVideo';
